requirements.txt
asgiref==3.6.0
cssselect==1.2.0
Django==3.2
django-allauth==0.41.0
django-crispy-forms==1.14.0
django-jsignature==0.11
lxml==4.9.2
oauthlib==3.2.2
Pillow==9.4.0
pyquery==2.0.0
python3-openid==3.2.0
pytz==2022.7.1
requests-oauthlib==1.3.1
sqlparse==0.4.3
